requirements.txt
# local package
-e .

# external requirements
click
python-dotenv>=0.5.1
tqdm
joblib
numpy
pandas
matplotlib
seaborn
lightgbm
scikit-learn
xgboost